translation_stm: reset highest translated offset on raft snapshot
When an stm receives Raft snapshot it indicates the whole in memory
state of that state machine should be replaced with the state from the
snapshot. The datalake translation state machine was incorrectly
handling raft snapshot which lead to its state being out of date after
the snapshot is applied. Raft snapshot for translation_stm is empty so
the correct action is to reset the state machine state and wait for the
update to be applied.

Fixes: CORE-8485

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 8985f5762f42b0326e0a3f444b1d18cd1e0de8a3)