tx/groups: escape hatch for unsafe aborting of group transactions
This is unsafe because it does not do any required checks to see
if a particular transaction is in progress and is a candidate for abort.
For example if a transaction is committed by the coordinator and
pending commit on the group, using this escape hatch to abort the
transaction can cause correctness issues. To be used with caution as an
escape hatch for aborting transactions that the group has lost track of
are ok to be aborted. This situation usually is indicative of a bug in
the transaction implementation.

(cherry picked from commit 8c5ecca7b592d4ddc9b1f2c61a8ebc89d153604f)