rm_stm: fix incorrect eviction attempts using a stale pid
pid is captured in the lambda could be a stale if the pid got fenced
(with an epoch bump). The change forces to provide a pid as a part of
the eviction hook, which would be the current pid at the time of
eviction.

(cherry picked from commit 79362faf97fa59a675e4edd090851c752a05b9aa)