archival: Handle nested shutdown errors
Nested exceptions are treated as shutdown errors only if both inner and
outer exceptions are shutdown errors. But sometimes the connections are
closed during the shutdown and the disconnect error is triggered. In
this case the exception is reported as true error. This increases the
noise level and complicates log monitoring.

This commit changes nested exception handling. If one of the exceptions
is a shutdown error and another one is a disconnect exception (broken
pipe) the error is treated as a shutdown error.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit 01ee010465a1bfeb9b9dfbdee134a09de424bba6)